Solved error in Firebase initialization
