+++ conf/application.conf
@@ -9,4 +9,3 @@ play.filters.hosts {
   # Allow requests to example.com, its subdomains, and localhost:9000.
   allowed = ["."]
 }
-
