--- conf/firebase.scala
@@ -1,7 +1,0 @@
-import java.io.InputStream
-
-import com.google.firebase.{FirebaseApp, FirebaseOptions}
-
-Object Firebase {
-
-}
